+++ tgui/packages/common/math.ts
@@ -45,7 +45,7 @@ export const round = (value, precision) => {
   m = Math.pow(10, precision);
   value *= m;
   // sign of the number
-  sgn = (value > 0) | -(value < 0);
+  sgn = +(value > 0) | -(value < 0);
   // isHalf = value % 1 === 0.5 * sgn;
   isHalf = Math.abs(value % 1) >= 0.4999999999854481;
   f = Math.floor(value);
