+++ tgui/packages/tgui/components/Flex.tsx
@@ -78,7 +78,7 @@ export const computeFlexItemProps = (props: FlexItemProps) => {
     className: classes([
       'Flex__item',
       Byond.IS_LTE_IE10 && 'Flex__item--iefix',
-      Byond.IS_LTE_IE10 && grow > 0 && 'Flex__item--iefix--grow',
+      Byond.IS_LTE_IE10 && (grow && grow > 0) && 'Flex__item--iefix--grow',
       className,
     ]),
     style: {
