+++ tgui/tsconfig.json
@@ -14,6 +14,7 @@
     "resolveJsonModule": true,
     "skipLibCheck": true,
     "strict": false,
+    "strictNullChecks": true,
     "target": "ES3"
   },
   "include": [
